requirements.txt
tk
nfcpy
pyserial
google-api-python-client
google-auth-oauthlib
pygame